backend/requirement.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
fastapi==0.115.11
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.70.0
pydantic==2.11.1
pydantic_core==2.33.0
sniffio==1.3.1
starlette==0.46.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0

twilio
pymongo==4.6.3
phonenumbers==8.13.33
python-dotenv==1.0.1
python-multipart
google-cloud-storage
uvicorn[standard]==0.34.0   

# CPU-only PyTorch stack (no CUDA, saves space)
torch==2.0.1+cpu
torchvision==0.15.2+cpu
torchaudio==2.0.2+cpu
silero-vad>=4.0.0
numpy==1.26.4
pydub

# Prometheus monitoring
prometheus-fastapi-instrumentator
prometheus-client
starlette
